Make Date::Manip an optional module. This involved:
A simple and permissive ISO-8601 date parser in lib/Astro/Coord/ECI/Utils.pm.

A new test, t/parse_time.t to make sure it works as advertised.

Changes in bin/satpass to use Date::Manip if it is available, but the ISO-8601
parser if not. These changes are pretty ugly, but hopefully that will be more
incentive to get on with App::Satpass2.

Machinery to test without optional modules. These are run as part of the
authortest, and then only if Test::Without::Module or Devel::Hide (in
that order) can be loaded. This of course involved
  Makefile.PL
  MANIFEST
  inc/Astro/Coord/ECI/Build.pm
  Build.PL
  MANIFEST.SKIP

The optional-module machinery creates xt/generated (which svn had to be
taught to ignore) and populates it with wrappers for all the tests in t/.
These wrappers hide the optional modules and then 'require' the
appropriate file from t/.

Making Date::Manip a recommended module rather than a required module,
which also involved Makefile.PL and Build.PL.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-Coord-ECI/trunk@1521 f6fd3633-1913-47c2-bad4-372d75c948ec