Decided that ignoring Time::y2038 on 64-bit-integer systems was a bit too precipitate, so I use it anyway, but modify the message generated by inc/Astro/Coord/ECI/Recommend.pm to express my belief that it is unnecessary.
But all the logic on which versions of timegm, gmtime and friends to use
has been centralized to Astro::Coord::ECI::Utils, which exports whatever
it decides on to the other modules. So if I change my mind yet again, I
only need to modify one place.

Note that t/parse_time.t still needs to know which module is in use,
because if Time::y2038 has overridden the time modules in
Astro::Coord::ECI::Utils, this test needs to re-override them with the
Test::MockTime routines, since Test::MockTime overrides in CORE::GLOBAL,
but Time::y2038 overrides locally.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-Coord-ECI/trunk@1571 f6fd3633-1913-47c2-bad4-372d75c948ec