Add Astro::Coord::ECI::TLE 'rcs' attribute, to hold radar cross-section.
Add the ability of Astro::Coord::ECI::TLE->parse() to parse --rcs num
off the the name line of a NASA tle, the way it did --effective date.

Document the Astro::Coord::ECI::TLE 'effective' attribute (ooops!).

Test the parsing of of the --effective and --rcs ad-hoc data off the
name line of a NASA-format tle.

Add 'rcs' as a stop word in xt/author/pod_spelling.t.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-Coord-ECI/trunk@1915 f6fd3633-1913-47c2-bad4-372d75c948ec