In t/tle_pass.t:
Add Epsilon Leonis to the test of passes of 88888 over Greenwich
England, so I get a test of the appulse logic.

Add tests of relevant passes over Holland and Bogota Columbia to test
the corner cases uncovered by Jaap Meijers when converting his Twisst
twitter service from Heavens Above-scraping to Astro::Coord::ECI::TLE.
These use real live ISS elements fetched from Space Track, so they have
to be in a SKIP block, contingent on the presence of the file the TLE
are in, since I can't distribute the file.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-Coord-ECI/trunk@2372 f6fd3633-1913-47c2-bad4-372d75c948ec