All the following in the Astro::Coord::ECI::TLE pass() method:
* Redo support for the Astro::Coord::ECI::TLE 'interval' attribute. The
  new implementation throws away the rough calculations and recalculates
  from scratch. This should be much faster, though Your Milage May Vary.

* Fix a regression in the code that calculates the max altitude of the
  satellite. This was introduced in 0.039_03 when I fixed problem of
  missing passes close to the end of the prediction interval.

* Extend the previous fix for passes in progress at the start of the
  prediction interval to predictions made with the 'visible' attribute
  false.

Split out the optional tests from t/tle_pass.t into t/tle_pass_extra.t.
Consolidate all TLE data for this into pass_extra.tle, which is expected
in the directory pointed to by environment variable
ASTRO_COORD_ECI_TLE_DIR. In the absence of that environment variable,
try to load File::HomeDir and look for it in the my_dist_config()
directory for Astro-Coord-ECI-TLE-Dir. The data are stored outside the
distribution so users can set it up once and for all. The contents of
pass_extra.tle are validated, so the user knows when changes are made.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-Coord-ECI/trunk@2404 f6fd3633-1913-47c2-bad4-372d75c948ec