Replace the Astro::Coord::ECI::TLE 'limb' attribute and the satpass 'lit' setting with the 'edge_of_earths_shadow' attribute and setting. The latter is numeric, and is intended to increase (or decrease if negative) the elevation above the horizon by the given number of radiuses of the body in question (typically the illuminating body). The former are now deprecated, and implemented in terms of the latter.
As for the implementation itself, this meant splitting out
Astro::Coord::ECI method azel_offset(), where the optional second
argument is the offset of the elevation in terms of apparent radii of
the body whose position is being calculated. The azel() method is
implemented in terms of azel_offset(), with the second argument (if
provided) being mapped into 1 or 0, according to whether it is true or
false. The two-argument version of azel() is now deprecated.

Also, overrode Astro::Coord::ECI method correct_for_refraction() in
Astro::Coord::ECI::TLE, Astro::Coord::ECI::Moon and Astro::Coord::Sun,
to simply return the original elevation. This because whatever the
atmospheric refraction is in this case, it is _not_ what is implemented
in Astro::Coord::ECI.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-Coord-ECI/trunk@2508 f6fd3633-1913-47c2-bad4-372d75c948ec