Add TLE method __list_type() (experimental)
UNSUPPORTED AND EXPERIMENTAL, at least at this time.

Add method __list_type() to Astro::Coord::ECI::TLE. This returns
'inertial' if the coordinates are inertial, and 'fixed' if not. The
intent is that Astro::App::Satpass2 should be able to use this to pick,
on a per-object basis, a template to format the object in the list()
command.