Changes to support Astro::Coord::ECI::Points
All the following are in Astro::Coord::ECI::TLE:

Add method intrinsic_events(), which computes events intrinsic to the
class over the pass period. This is called by pass() once it has a pass.
This version returns nothing, but gives a hook for a subclass to compute
pass events over and above the ones computed by pass(). This method is
to be considered experimental.

Add protected method __pass_backup_earliest(), which returns 0. This is
called by the pass() logic which tries to compute the rise time of a
pass that was already in progress when the computation interval started.
This was needed because the Astro::Coord::ECI::Points algorithm (with
good reason) considers the body at the start position any time before
the start time, so without this the computation is an endless loop if
the launch pad is above the viewer's horizon. This method is to be
considered unsupported and subject to change or retraction without
notice.

Centralized the computation of illumination changes (or not, depending
on object configuration) into private method _find_illumination().

Centralized the computation of periodic positions (or not, depending on
object configuration) into private method _find_position().