Fix problem with objects wrongly becoming inertial
This was complicated. The problem was triggered by having the
equinox_dynamical attribuite set on the observing station and then doing
a pass() calculation in which the station was an attribute of the TLE
and appulse calculations to a star needed to be made.

To cut to the chase, the Star object precesses its own coordinates to
the relevant meridian when the time is set. The precess_dynamical()
method checks for the presence of the 'station' attribute, and precesses
that too. But precess_dynamical() ends up calling equatorial() to set
the coordinates. This makes the object inertial, whether it originally
was or not.

The changes are in Astro::Coord::ECI->precess_dynamical(), and there are
two:

1) Do not precess the coordinates of Earth-fixed stations.

2) If we do (for some reason) call precess_dynamical() on an Earth-fixed
object, retrieve the original coordinate system, and after setting the
object's coordinates in Equatorial coordinates, retrieve and re-set them
in the original coordinates. Since the original coordinates are
Earth-fixed, a side effect of this is to make the object Earth-fixed
again.

Either of the above makes the problem go away. I only did (2) because
someone might really want to precess Earth-fixed data, and the
alternative was to throw an exception.