Work around JSON::XS locale problem.
The problem may be Perl rather than JSON::XS, but the latter is the way
the discussion on the Perl ticket is leaning. The problem is leakage of
locale information into the encoded JSON if JSON::XS does the encoding.
If the locale uses commas for decimal points you get invalid JSON.

The thing is that I can't really work around this in the code, because
of the way the JSON interface to custom objects works. What I have done
is:

1) Set LC_NUMERIC to "C" in t/tle_json.t, since this was where the
problem manifested, and just to get the test to pase.

2) Added a note about the situation in Astro::Coord::ECI::TLE.

The problem is that I just pass a hash reference to JSON, and it takes
care of stringifying it. So I have no control over the actual
stringification. Sigh.