Prevent TLE manufacture for oddball objects
What this is really about is trying disable the manufacture of a 'tle'
attribute value from the other attributes unless it has a reasonable
chance of success.

While I was at it, cleaned up the exception message thrown when one of
the attributes required to manufacture a TLE is undefined.