More correct error checking when making TLEs
The issue here is that there is code to generate TLE from objects that
were initialized by setting their attributes directly rather than by
parsing a TLE. But (probably due to design problems that have become too
deeply embedded to be easily changed) there are Astro::Coord::ECI
objects that legitimately do not have associated TLE data, and for which
the attempt should return nothing rather than generate an error.

Astro::Coord::ECI::TLE v064_02 attempted to solve this by checking
whether the object was inertial or not. But this turned up a problem in
the Astro-App-Satpass2 test suite. So the current try is to simply count
the required fields which are null. If they all are, we simply return
nothing. If at least one is not, than we assume an unsuccessful attempt
to generate a TLE and report any missing fields.