Have Iridium can_flare( 'all' ) return true.
This is regardless of status, and anticipated as being useful once they
all go out of service. The superclass' can_flare() still returns false
unconditionally.

Also needed to upgrade support of decayed bodies in status( add ... ). I
may want to rethink this, since I tripped over the fact that I was
keeping a decoding table in script/satpass. This really needs to be in
::Iridium.