Correct typo in calculation of Sun's postion
The typo was in the calculation of $omega, which Meeus does not define
but which I believe to be the ascending node of the Sun. The correction
was the coefficient of $T (Julian century) from 1934.156 to 1934.136.
This value is used in adjusting the geometric longitude of the Sun to
the apparant longitude of the Sun, and produced no change at all in the
output of eg/usno-seasons for 2018. That it, I got the same season times
before and after the change (to the nearest second), and the same
calculated longitude at the U.S. Naval Observatory's season times (to 4
decimal places) whether or not the change was in place.