t/tle_support.t now checks ::TLE::Iridium version
Since this was pulled into its own distro we can't count on having the
latest version. But we can't drop this test either because we need to be
sure we're delivering the support Astro::Coord::ECI::TLE::Iridium needs.
So one test got conditionalized on version number. Are there more? Who
knows?