Fix $eci->set( sun => ... )
The mutator was not in fact setting the new value of the attribute.

Included in this is support for changing the static (i.e. default) value
of the attribute. This in turn meant cloning the static value if
singletons were not being used. I decided that having clone() return a
new object if singletons were being used was a bug, and overrode the
clone method for both Sun and Moon.