Simplify sprintf code re 'no warnings redundant'
This was already run through a package-private subroutine, __sprintf().
The issue is that the 'redundant' warnings category was introduced in
5.21.2, but you had to turn it off to get the pre-5.21.2 behavior. This
manifested because I did error reporting by passing a format and some
arguments, without requiring the format to use all the arguments.

The previous implementation was to stringy-eval the subroutine into
existence, with the 'no warnings qw{ redundant }' included or excluded
based on $].

The current implementation puts the code in line, and disables the
warning if necessary using
  no if $] ge '5.021002' qw{ warnings redundant };