Fix broken POD links.
These were mostly in script/satpass, which does not get checked unless
you actually request it to be installed.

The "Ask Dr. Math" link on haversines is also gone. It appears the
teachers' association that took it over from Drexel decided non-members
were unworthy of this knowledge. Another group is trying to recover them
from the archives, but if you need something you have to request it. I
have done so, and we'll see what happens.