Replace my_strftime() with gm_strftime() and local_strftime().
The replacements take a strftime() template, an epoch, and an optional
number of decimal places for seconds. The template can also specify a
default number of decimal places for seconds for any item that
represents seconds. This extension is analogous to strftime(): the
specification is (eg) '%.2S' for seconds to hundredths. If the template
specifies decimal places, it overrides the argument.