Add method next_elevation_extreme_tod().
This is analogous to next_elevation(), but returns the earliest (or
latest) time the body passes the given elevation after the currentlt-set
time -- or nothing if there is none. It really is only useful for
computing earliest or latest Sunrise or Sunset, but is present on
Astro::Coord::ECI for architectural reasons (meaning my architecture is
flawed, but that can't be changed now).